# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Heures UVCI API",
        "description": "Documentation de l'API de gestion des heures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/v1/activites": {
            "get": {
                "tags": [
                    "Activités"
                ],
                "summary": "Liste des activités",
                "operationId": "0cfa867fdaebb196dc2e4562ee321012",
                "parameters": [
                    {
                        "name": "id_enseignant",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_annee",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des activités"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activités"
                ],
                "summary": "Enregistrer une activité pédagogique",
                "operationId": "156b46a5b8433fb9c754f8ff29dd2512",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_ressource",
                                    "nb_seances",
                                    "date_activite",
                                    "id_attribution",
                                    "id_annee"
                                ],
                                "properties": {
                                    "id_ressource": {
                                        "type": "integer"
                                    },
                                    "nb_seances": {
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2,
                                            4,
                                            6
                                        ]
                                    },
                                    "date_activite": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "id_attribution": {
                                        "type": "integer"
                                    },
                                    "id_annee": {
                                        "type": "integer"
                                    },
                                    "observations": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Activité enregistrée avec volume horaire calculé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}